fix(build): fix build errors
