+++ .env
@@ -1,6 +1,5 @@
 DATABASE_URL="file:./dev.db"
-NEXTAUTH_URL=""
-NEXTAUTH_URL_INTERNAL=""
+# NEXTAUTH_URL_INTERNAL=""
 GITHUB_ID=5bc082904c38ef2b7191
 GITHUB_SECRET=096ffbb24529d51ec20b626febf833640be4feeb
 NEXTAUTH_URL=http://localhost:3000
