+++ app/api/signup/helpers.ts
@@ -1,12 +1,6 @@
 import { NextResponse } from 'next/server';
 import { prisma } from '@/lib';
 
-// interface Details {
-//     theme: number | undefined;
-//     role: number | undefined;
-//     language: number | undefined;
-// }
-
 export const conectionObject = (value: string | number | undefined) => {
   return {
     connect: {
