+++ app/components/navbar/Logo.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 import { useContext } from 'react';
 import { useRouter } from 'next/navigation';
 
-import { ThemeContext } from '@/app/providers';
+import { ThemeContext } from '@/app/components/providers';
 
 // @ts-ignore
 const Logo = (props) => {
