+++ app/components/providers/auth.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { SessionProvider } from 'next-auth/react';
 
 // @ts-ignore
